// src/_gen/grpc/events/events.ts
export interface TicketSoldEvt {
  ticketId: string;
  userId: string;
  showtimeId: string;
  /** string test = 5; */
  price: number;
}

function createBaseTicketSoldEvt(): TicketSoldEvt {
  return { ticketId: "", userId: "", showtimeId: "", price: 0 };
}

export const TicketSoldEvt: MessageFns<TicketSoldEvt> = {
  encode(message: TicketSoldEvt, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ticketId !== "") {
      writer.uint32(10).string(message.ticketId);
    }
    if (message.userId !== "") {
      writer.uint32(18).string(message.userId);
    }
    if (message.showtimeId !== "") {
      writer.uint32(26).string(message.showtimeId);
    }
    if (message.price !== 0) {
      writer.uint32(33).double(message.price);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): TicketSoldEvt {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseTicketSoldEvt();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.ticketId = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.userId = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.showtimeId = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 33) {
            break;
          }

          message.price = reader.double();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): TicketSoldEvt {
    return {
      ticketId: isSet(object.ticketId)
        ? globalThis.String(object.ticketId)
        : isSet(object.ticket_id)
        ? globalThis.String(object.ticket_id)
        : "",
      userId: isSet(object.userId)
        ? globalThis.String(object.userId)
        : isSet(object.user_id)
        ? globalThis.String(object.user_id)
        : "",
      showtimeId: isSet(object.showtimeId)
        ? globalThis.String(object.showtimeId)
        : isSet(object.showtime_id)
        ? globalThis.String(object.showtime_id)
        : "",
      price: isSet(object.price) ? globalThis.Number(object.price) : 0,
    };
  },

  toJSON(message: TicketSoldEvt): unknown {
    const obj: any = {};
    if (message.ticketId !== "") {
      obj.ticketId = message.ticketId;
    }
    if (message.userId !== "") {
      obj.userId = message.userId;
    }
    if (message.showtimeId !== "") {
      obj.showtimeId = message.showtimeId;
    }
    if (message.price !== 0) {
      obj.price = message.price;
    }
    return obj;
  },

  create(base?: DeepPartial<TicketSoldEvt>): TicketSoldEvt {
    return TicketSoldEvt.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<TicketSoldEvt>): TicketSoldEvt {
    const message = createBaseTicketSoldEvt();
    message.ticketId = object.ticketId ?? "";
    message.userId = object.userId ?? "";
    message.showtimeId = object.showtimeId ?? "";
    message.price = object.price ?? 0;
    return message;
  },
};
// ...
export interface GatewayEventsClient<CallOptionsExt = {}> {
  /** Событие продажи билета */
  ticketSold(request: DeepPartial<TicketSoldEvt>, options?: CallOptions & CallOptionsExt): Promise<Empty>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}